% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHSInD.R
\name{LHSD}
\alias{LHSD}
\title{Latin Hypercube Sampling in domain D}
\usage{
LHSD(N, D)
}
\arguments{
\item{N}{The number of samples to generate.}

\item{D}{A matrix where each row represents the range min, max for a design variable.}
}
\value{
A matrix of size N x d, where d is the number of design variables (rows of D).
}
\description{
Generates a Latin Hypercube Sample (LHS) scaled to fit within the domain.
}
\examples{
D <- matrix(c(0, 10, 0, 5), ncol = 2, byrow = TRUE)
LHSD(100, D)
}
