% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbor.R
\name{Neighbor}
\alias{Neighbor}
\title{Generate Random Neighboring Points}
\usage{
Neighbor(x, D)
}
\arguments{
\item{x}{A numeric vector representing the initial point.}

\item{D}{A numeric matrix with two columns. Each row represents the minimum and maximum bounds for the corresponding element of \code{x}.}
}
\value{
A numeric vector of the same length as \code{x}, containing the generated neighboring points.
}
\description{
This function generates a vector of random neighboring points around a given point \code{x}, constrained by the range specified in matrix \code{D}.
}
\examples{
x <- c(1, 2)
D <- matrix(c(0, 3, 1, 4), ncol = 2, byrow = TRUE)
Neighbor(x, D)
}
