% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{SAoptim}
\alias{SAoptim}
\title{Simulated Annealing Method for Optimal Design of Experiments}
\usage{
SAoptim(
  fobj,
  size,
  D,
  maxiter = NULL,
  temp = 0.1,
  c = 0.99,
  Xinit = NULL,
  ptype = "metropolis",
  tempscheme = "geo",
  isMax = TRUE
)
}
\arguments{
\item{fobj}{A function representing the objective function to be optimized. It should accept a matrix as input where each row is a different set of experiments.}

\item{size}{An integer specifying the size of the experimental design (number of points).}

\item{D}{A numeric matrix representing the domain of design elements. Each row defines the range for one design variable.}

\item{maxiter}{Maximum number of iterations for the algorithm (default: 1000).}

\item{temp}{Initial temperature for the annealing process (default: 1).}

\item{c}{Cooling rate, controlling the reduction of temperature over iterations (default: 0.99).}

\item{Xinit}{Initial design matrix for the experiments. If NULL, the initial design from \code{object} is used.}

\item{ptype}{The probability scheme used for accepting new designs. Can be "metropolis" (default) or "glauber".}

\item{tempscheme}{The temperature scheme used to reduce the temperature. Can be "geo" for geometric (default) or "log" for logarithmic.}

\item{isMax}{The parameter that specifies whether the function is to be maximized or minimized. Its default value is "TRUE".}
}
\value{
A list with the optimal design (\code{Xopt}) and the corresponding criterion value (\code{Copt}) found.
}
\description{
This function optimizes a criterion using the Simulated Annealing (SA) method. It iteratively searches for the best experimental design within a specified domain by probabilistically exploring neighboring designs.
}
\examples{
fobj <- function(X){return(log(det(t(X)\%*\%X)))}
D <- matrix(c(0, 10, 0, 5), ncol = 2, byrow = TRUE)
R <- SAoptim(fobj, size=10, D=D, maxiter = 1000, temp = 1, c = 0.95)
print(R$Xopt)
print(R$Copt)
}
