% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckInD.R
\name{TT}
\alias{TT}
\title{Constraint Handling Function for Optimization}
\usage{
TT(X, D)
}
\arguments{
\item{X}{A numeric matrix where each row represents a candidate solution in the optimization process.}

\item{D}{A numeric matrix with two columns wich represent the domain.}
}
\value{
A matrix \code{X} with the same dimensions as the input, but with all values
constrained within the bounds defined by \code{D}.
}
\description{
The \code{TT} function ensures that all values in a matrix \code{X} are constrained
within specified bounds defined by a matrix \code{D}.
}
\details{
This function iterates through each element of the matrix \code{X}. For each element,
it checks if the value is within the bounds specified by the corresponding row in \code{D}.
If the value is below the lower bound, it is replaced by the lower bound.
If it is above the upper bound, it is replaced by the upper bound.
}
\examples{
X <- matrix(runif(10), nrow=5, ncol=2)
D <- matrix(c(0.2, 0.8, 0.1, 0.9), nrow=2, ncol=2, byrow=TRUE)  # Bounds
Xc <- TT(X, D)
print(Xc)

}
